README.md
The Restaurant dels Solitaris is famous for being the only one that only has individual tables and where you can still smoke. The restaurant is divided into X rooms with N tables each. As there have been complaints from some customers about smoking by smokers, the owner of the restaurant has decided to establish the following protocol: When a customer arrives, he asks the waiter for a table, indicating whether he is a smoker or not, and the waiter will try to give him a table according to the following criteria:
- When a room is empty, both smokers and non-smokers can enter.
- Smoking customers can only enter smoking rooms.
- As soon as the first customer (smoker/non-smoker) enters a room, the use of the room is restricted to this type of diner. This will be the case until the room is empty.
Once the customer has dined, they ask for the bill from the waiter who takes into account the table that has been empty.